chicky_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
openai==1.51.0
anthropic==0.34.2
edge-tts==6.1.12
python-multipart==0.0.9
websockets==13.1
python-dotenv==1.0.1
supabase==2.7.4
aiofiles==24.1.0
